+++ src/test/java/org/icgc/argo/program_service/AssociatorTest.java
@@ -25,7 +25,7 @@ import org.icgc.argo.program_service.model.entity.PrimarySiteEntity;
 import org.icgc.argo.program_service.model.entity.ProgramEntity;
 import org.icgc.argo.program_service.model.join.ProgramCancer;
 import org.icgc.argo.program_service.model.join.ProgramPrimarySite;
-import org.icgc.argo.program_service.relationship.Associator;
+import org.icgc.argo.program_service.association.Associator;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
@@ -37,8 +37,8 @@ import java.util.function.Supplier;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.icgc.argo.program_service.RandomGenerator.createRandomGenerator;
-import static org.icgc.argo.program_service.relationship.Associators.PROGRAM_CANCER_ASSOCIATOR;
-import static org.icgc.argo.program_service.relationship.Associators.PROGRAM_PRIMARY_SITE_ASSOCIATOR;
+import static org.icgc.argo.program_service.association.Associators.PROGRAM_CANCER_ASSOCIATOR;
+import static org.icgc.argo.program_service.association.Associators.PROGRAM_PRIMARY_SITE_ASSOCIATOR;
 import static org.icgc.argo.program_service.utils.CollectionUtils.mapToImmutableList;
 import static org.icgc.argo.program_service.utils.CollectionUtils.repeatedCallsOf;
